SSI7X/requirements.txt
Flask==0.12.2
Flask_Restful==0.3.6
Flask_Cors==3.0.3
psycopg2==2.7.3.1
WTForms==2.1
IPy==0.83
ldap3==2.4
ua-parser==0.7.3
user-agents==1.1.0
Flask-Session==0.3.1
PyJWT==1.5.3
XlsxWriter==1.0.4
numpy==1.14.3
pandas==0.22.0